admin routes protected except for all orders page

+++ server/controllers/productsController.js
@@ -20,6 +20,19 @@ const getProductById = async (req, res, next) => {
   }
 };
 
+const getProductByGender = async (req, res, next) => {
+  try {
+    const { gender } = req.params;
+    const filteredProducts = await Product.find({
+      "features.gender": gender,
+    });
+    console.log(filteredProducts);
+    res.status(200).json(filteredProducts);
+  } catch (err) {
+    res.status(500).json(err);
+  }
+};
+
 const createProduct = async (req, res, next) => {
   //Create Product
   try {
@@ -35,30 +48,21 @@ const createProduct = async (req, res, next) => {
       },
     ];
     const { brand, price, onSale, description, image, features } = req.body;
-    const product = await Product.create({
-      brand,
-      price,
-      onSale,
-      description,
-      image,
-      colors,
-      pickedQty: 1,
-      features,
-    });
-    res.status(200).json(product);
-  } catch (err) {
-    res.status(500).json(err);
-  }
-};
-
-const getProductByGender = async (req, res, next) => {
-  try {
-    const { gender } = req.params;
-    const filteredProducts = await Product.find({
-      "features.gender": gender,
-    });
-    console.log(filteredProducts);
-    res.status(200).json(filteredProducts);
+    if (req.admin) {
+      const product = await Product.create({
+        brand,
+        price,
+        onSale,
+        description,
+        image,
+        colors,
+        pickedQty: 1,
+        features,
+      });
+      res.status(200).json(product);
+    } else {
+      res.status(400).json("You are not authorized to do this.");
+    }
   } catch (err) {
     res.status(500).json(err);
   }
@@ -68,14 +72,18 @@ const updateProduct = async (req, res, next) => {
   try {
     const productId = req.params.id;
     const { brand, price, onSale, description, image, features } = req.body;
-    const updateProduct = await Product.findByIdAndUpdate(productId, {
-      brand,
-      price,
-      description,
-      image,
-      features,
-    });
-    res.status(200).json(updateProduct);
+    if (req.admin) {
+      const updateProduct = await Product.findByIdAndUpdate(productId, {
+        brand,
+        price,
+        description,
+        image,
+        features,
+      });
+      res.status(200).json(updateProduct);
+    } else {
+      res.status(400).json("You are not authorized to to this.");
+    }
   } catch (err) {
     res.status(400).json(err);
   }
@@ -83,10 +91,15 @@ const updateProduct = async (req, res, next) => {
 
 const deleteProduct = async (req, res, next) => {
   try {
-    const productId = req.params.id;
-    const deletedProduct = await Product.findByIdAndRemove(productId);
+    if (req.admin) {
+      const productId = req.params.id;
+      const deletedProduct = await Product.findByIdAndRemove(productId);
+    } else {
+      res.status(400).json("You are not authorized to do this.");
+    }
     res.status(200).json(deletedProduct);
   } catch (err) {
+    console.log(err);
     res.status(400).json(err);
   }
 };
